requirements.txt
streamlit==1.37.0
python-pptx==0.6.23
requests>=2.31.0
Pillow>=10.0.0       # For image handling in PPT
openai>=1.0.0        # If you support OpenAI API
anthropic>=0.3.0     # If you support Anthropic API
google-generativeai  # For Gemini API
